Introduce domain package
