+++ semantics/services/src/main/java/net/catenax/semantics/hub/domain/ModelsPackage.java
@@ -13,9 +13,16 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package net.catenax.semantics.hub.persistence.triplestore;
+package net.catenax.semantics.hub.domain;
 
-public enum ModelsPackageStatus {
-   RELEASED,
-   DRAFT
+public class ModelsPackage {
+   private final ModelsPackageStatus status;
+
+   public ModelsPackage( ModelsPackageStatus status ) {
+      this.status = status;
+   }
+
+   public ModelsPackageStatus getStatus() {
+      return status;
+   }
 }
